Better sync of encoder and target position timing
Prior to this change, extra noise was introduced in the error signal
because some non-negligible time could elapse between when the target
position was calculated and when the mount encoders were read. This time
could also vary depending on how long the intervening processing took.
Now the target position is read for a time that corresponds as closely
as possible to when the encoder positions were read from the mount.