Experiments with blind error time synchronization
It was discovered that a significant source of noise in the blind error
source is the difference between the instant in time when the mount
encoders are read and the time that is assumed when calculating the
target position. In this commit, an experiment was performed to assess
whether the execution time of the code that runs between when the
network message with the encoder positions is received from Gemini and
the mount object's get_position() method returns is significant. These
experiments showed the following:

* Elapsed time between the events described is about 0.3 ms on average
* The jitter in this time delta is significant compared to the mean;
spikes up to ~1 ms were observed.
* No correlation between these spikes in the time delta and the blind
error magnitude were found.

Note that this commit depends on corresponding changes in the point
package; specifically, the enq_macro() method must return a timestamp
along with the normal result, and the timestamp must be recorded
immediately after the call to recv() in the backend.