#177: Optimize target offset calc CPU load
The target position offset calculations used by the blind error source
are now only applied when the offset is non-zero and when the B button
on the gamepad is actively held down. This is to prevent unintentional
application of the offset and also to eliminate the CPU time consumed by
these calculations when they are unneeded.

Also optimizes the calculations by caching the last two target positions
in the blind error source object such that they are always available for
calculation of the target direction of travel when needed.