Separate camera args for alignment and tracking
The gain and exposure time settings for the camera are provided by
program arguments which may be stored in a configuration file. The
values for these parameters that are appropriate for alignment, which
need longer single-shot exposures, are not the same values appropriate
for tracking, which requires shorter exposures. This commit creates
separate arguments for both cases so they can be specified separately in
the same config file.