More intermediate transforms in MountModel
The `MountModel` class contains methods for converting between mount
encoder positions and topocentric coordinates. The intermediate step is
a coordinate in a mount-relative spherical coordinate system. This
commit adds the transformation between this intermediate representation
and topocentric coordinates as dedicated methods. In certain cases the
availability of these methods can reduce CPU load by bypassing the part
of the full transformation that is not needed.