program arg for guidescope cam orientation
Eventually it is hoped that alignment will include solving for the
orientation of the guidescope camera (#149) but this has not been
implemented. This does the next-best thing, which is to replace the
hard-coded value with a program argument so that it can at least be set
at the command line or in a config file.