#187: Alternatives when stored alignment is stale
In some situations a fresh set of model parameters from alignment are
not necessary. Prior to this commit, if the model parameters were
stale the program would just abort. Now a prompt allows for using the
stale parameters anyway or, in the case of optical tracking, falling
back on a default set of model parameters.