Revert gamepad_control callback
The callback approach applied in gamepad_control.py in the previous
commit was broken because it did not take into account that slew()
commands should be sent when the integrator thread updates the
integrator values and these do not correspond to any gamepad events.
Thus this approach was reverted. However to improve responsiveness it
was found that the multiprocessing feature recently added to the Gemini
mount class was unnecessary for this program and was in fact a major
contributor to the responsiveness issues. By disabling this feature the
slew commands are sent synchronously with the loop in the main thread.