Un-break gamepad control with Losmandy mount
This fixes an unintended change in behavior that was introduced in
commit 0f1b6c1f. Prior to 0f1b6c1f, starting in 833c9ef7, the
constructor to the LosmandyGeminiMount class was called with
use_multiprocessing set to False. This is required for smooth behavior
with gamepad control. In 0f1b6c1f some refactoring of how mounts are
constructed across all programs left this constructor argument at the
default value when is True.