Adds an API for accessing mount axis limits
I don't recall exactly why this was needed but it was in a stash. It may
have something to do with looking ahead to ensure that the route the
control system is taking the mount won't hit a limit and then require a
lengthy slew after crossing over the pole.